// src/generated/models/Cgmp_blackoutperiodsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cgmp_blackoutperiodsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cgmp_blackoutperiodsstatecode = keyof typeof Cgmp_blackoutperiodsstatecode;
export const Cgmp_blackoutperiodsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cgmp_blackoutperiodsstatuscode = keyof typeof Cgmp_blackoutperiodsstatuscode;

export interface Cgmp_blackoutperiodsBase {
  cgmp_affectedlocations?: string;
  cgmp_blackoutperiodid: string;
  cgmp_createdbyupn?: string;
  cgmp_enddate: string;
  cgmp_name?: string;
  cgmp_reason?: string;
  cgmp_startdate: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cgmp_blackoutperiodsstatecode;
  statuscode?: Cgmp_blackoutperiodsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cgmp_blackoutperiods extends Cgmp_blackoutperiodsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}